add rough work, switching class manip functions to classes
