+++ menelaus/injection/__init__.py
@@ -4,10 +4,11 @@ from menelaus.injection.feature_manipulation import (
     feature_swap,
     feature_hide_and_sample,
 )
-from menelaus.injection.class_manipulation import (
-    class_swap,
-    class_join,
-    class_dirichlet_shift,
-    class_probability_shift,
+from menelaus.injection.injector import Injector
+from menelaus.injection.label_manipulation import (
+    LabelSwapInjector,
+    LabelJoinInjector,
+    LabelProbabilityInjector,
+    LabelDirichletInjector,
 )
 from menelaus.injection.noise import brownian_noise, random_walk
